Fix several embedded pathnames that might cause problems on non-Unix machines.
Change clean_out_dir() to use dir_list() instead of list_dir(), as it was
not making any use of the extra features of list_dir. Also, as there was
one remaining use of list_dir(), which was the checks directory (which had
some special cases burned in), delete list_dir and replace it by a special
purpose function for that case.