Document the --no-asm option
Wrap lines a little tighter in the output of --help

Change the order of checks so we don't display that a module was not
used due to use of --no-asm when in fact the module would not run on
the target CPU anyway. Similiarly don't say "loaded only on request" if
it is incompatible with the target, since someone attempting to actually
load the module would immediately get an error.