Rename fips140.h to selftest.h, since it doesn't really have much to do with FIPS-140 aside from being a startup self test.
Remove the test of X9.19 MAC at startup because it's not that important, and
loading it in the startup means a prototypical object will be created and
exist in memory for the entire process runtime. This actually raises an
interesting idea, that periodically the cache of objects could be cleared
and, if one is needed again, it can be created again as if it was the first
time.