If an unbound variable in a template is found, print the all the currently existing variables and the first 60 characters of their values.
Simplify match_any_of - it just uses it's first two arguments and then
matches against the rest, rather than forcing them to be joined into a
single string (that it would just split and iterate over anyway).

Add support for two new compiler config vars, dll_import_flags and
dll_export_flags

When autoconfig prints that it has guessed something, provide the name
of the variable that can be used to override it.