Add BOTAN_DLL macro in all needed spots for working DLL export. Based on a patch from Joel Low for MSVC, modified and tested with GCC using -fvisibility=hidden and the visibility attribute.
Getting this working completely requires making the shared object and
static lib builds completely distinct (which is also a win since it
allows avoiding -fPIC usage, etc in the static lib).

Currently too many things are being exported, though it is an improvement
as internal-only code like the bigint_* functions are hidden.