Move more of the public key implementation code out of the shared pk/pubkey directory and into algorithm specific modules.
Even now, disabling a public key algorithm does leave some 'residual' code,
but much less than before.

Mark if_algo to only load if needed (if RW or RSA is loaded). However there seems
to be a bug in the dependency handling in configure.pl - even if rsa and rw
are disabled with --disable-modules, if_algo continues to be built. Will
have to look at this later.