Don't generate the makefile until after we have processed build.h.in
Markus Wanner noted on the mailing list that build.h was not being 
installed. That is because it was not included in the HEADERS variable
in the Makefile, and build.h did not yet exist when configure.pl scanned
the source and include directories.

Rearrange the order of operations in configure.pl a bit so first build.h
is generated and then the makefile is.