Add a new switch to configure.pl --with-tr1-implementation which accepts options 'boost' and 'system'. Now GF(p) math (and indirectly, ECDSA) will be enabled if --with-tr1=boost or --with-tr1=system is passed at build time to enable a shared_ptr implementation. Modules can now specify that they require TR1, in which case they will only be autoloaded if a TR1 implementation was set (they can choose which one to use by checking the feature macros from build.h)
The gfpmath module was set to load only on request. That has
changed to load automatically (but will only do so if a TR1
impl is set as described above). CVC has also been marked
as requiring TR1. (ECDSA/ECDH are not, since they do not
use tr1 shared_ptr directly)

Update and cleanup help output. Do not print the list of modules in
--help anymore (too long); you can still get the list (in an easier to
parse format) --module-info. Reorganize the help text so the more
useful options are described closer to the top.

Fix the --with-endian and --with-unaligned-mem options: they were being
accepted but ignored.