Fix the --with-build-dir option, so that it is actually possible to do out of tree builds.
Also rename the generated botan-config script so that it is, like the
pkg-config settings, namespaced by the major and minor version numbers
(eg, botan-17-config). This is useful in particular for distros like
Debian which ship both stable and unstable versions. Currently Debian
is actually the only distro I know of shipping 1.7 as well as 1.6, but
I would certainly like to encourage more in the future by making it
easy to do.