Add handy aliases to configure for controlling the modules using OpenSSL, GNU MP, zlib, and bzip2.
  --with-{openssl,gnump,bzip2,zlib}
  --without-{openssl,gnump,bzip2,zlib}

They have the exact same effect as --enable-modules=x or --disable-modules=x

This turned out to be a much easier way of specifying options for the
Gentoo ebuild. It is likely that other distro builds architectures will
also prefer this option style as being somewhat more autoconf-like and
fitting in with existing command templates.