In es_unix, two changes
Make the fast poll significantly more pessimistic/realistic about how
many bits of randomness we're getting from getrusage and stat.

Don't cut out from execing programs if the desired poll bits is under
128. Simply poll until either the accumulator says we're done or we run
out of sources. Assumption is that the poll won't be run at all unless
it is ncessary (es_unix comes late in the list of sources to use since
it is pretty slow).