CPU-specific engines are now only loaded if something depends on them, and all CPU-specific implementations now depend on the appropriate engine module.
The most common problem before with this was that the SSE2 module was built,
but the sole SSE2 code (SHA-1) was not (for instance, on an i686). This would
cause a compile warning about the unused request object.

Preventing unused engines from being built will also (very slightly) speed
up the lookup process on most system.