Minor changes to the license.
Remove the phrase 'for any use' - implied by the conditions.

Add 'All rights reserved.' after the (C) notice.

Remove the name+URL in the text and lead off with a statement that
Botan is distributed under the following terms: <blah>

Combine the two all-CAPS sentences disclaiming warranty into a single
paragraph.

The main reasoning behind this is to make the actual license text
totally equivalent to the FreeBSD/NetBSD licenses, which is an offical
Open Source(tm) license as declared by the OSI.