Add a log line telling the user that we are assuming unaligned memory accesses, if we are.
Add new options --with-unaligned-mem and --without-unaligned-mem to override
the arch-specific guess. The only usage I specifically know for this is to
allow unaligned accesses on m68k machines, when use on Coldfire is not
expected. Other cases where most CPUs of a particular type support unaligned
accesses but a specific target doesn't (or it is very slow) can also use this.