Increase the internal buffer size of the Hex coder/decoder, and put it into a named constant instead of being magic. Move from 64 bytes to 256.
This was necessary to allow
  Pipe(new Hex_Decoder, filter, ...)
to give filter a sufficiently large input block.

It would be nicer if the filter itself (in this case, ECB_Decryption, but
others apply as well) was smart enough to buffer on its own.

It might also be useful if code could query what parallelism a block cipher
provided and modify their actions accordingly.