Add a wrapper for a set of SSE2 operations with convenient syntax for 4x32 operations.
Also add a pure scalar code version.

Convert Serpent to use this new interface, and add an implementation of
XTEA in SIMD.

The wrappers plus the scalar version allow SIMD-ish code to work on all
platforms. This is often a win due to better ILP being visible to the
processor (as with the recent XTEA optimizations). Only real danger is
register starvation, mostly an issue on x86 these days. So it may (or may
not) be a win to consolidate the standard C++ versions and the SIMD versions
together.

Future work:
 - Add AltiVec/VMX version
 - Maybe also for ARM's NEON extension? Less pressing, I would think.
 - Convert SHA-1 code to use SIMD_32
 - Add XTEA SIMD decryption (currently only encrypt)
 - Change SSE2 engine to SIMD_engine
 - Modify configure.py to set BOTAN_TARGET_CPU_HAS_[SSE2|ALTIVEC|NEON|XXX] macros