Clean up the implementation of the line ending converter.
Add a custom install target to users can select specific things to enable
or disable.

Add api.pdf and tutorial.pdf, if they are available in the tree (otherwise skip)

Add the set of examples as part of the documentation package.

Require at least Windows 98 or 2000 since some code in the entropy
gathering routines requires functions (CryptGenRandom and Toolhelp32)
which are only available on these systems or later.

Set the VersionInfoVersion (viewable via the Properties menu on the setup .exe)