Prefix passhash with "$9$" in a manner similar with other password hashing schemes.
Increase salt size to 80 bits.

Research shows that virtually no other PBKDF2 implementations support
anything but SHA-1; for ease of implementation elsehwere switch back
from SHA-512 to SHA-1. Should be mostly harmless; it limits total
entropy of the password to a maximum of 160 bits, but this is unlikely
anyway.

Use two bytes to specify the work factor for future-proofing.

Add a test.