Further passhash changes before release and things have to be finalized.
Move header to passhash9.h and rename the functions to be passhash9
specific ({generator,check}_passhash9)

Add an algorithm identifer field. Currently only id 0 is defined, for
HMAC(SHA-1), but this opens up for using HMAC(SHA-512) or HMAC(SHA-3)
or CMAC(Blowfish) or whatever in the future if necessary. Increase the
salt size to 96 bits and the PRF output size to 192 bits.

Document in api.tex