Remove almost entirely unnecessary friend access from GFpElement to GFpModulus. Doing this pointed out what is probably a threading/race bug as well: assigning to a single GFpElement causes it to reset the GFpModulus to new values, but of course the other GFpElements don't know about this.
Sharing the GFpModulus was a really really bad design choice by the
InSiTo folks and it needs to die. That might end up wasting a bit of
memory, but it will probably fix a lot of nasty bugs, and also remove
the use of atomic operations which in the long run is going to hurt
performance a lot worse than some extra cache use.