Remove access to the Jacobian coordinate getters get_{x,y,z}, as well as the 4-argument constructor.
Define operator==() in terms of the affine coordinates.

Rewrite tests that assumed access to the Jacobian coodinates in terms
of the affine coordinates.

This change allows for using arbitrary coordinate systems in PointGFp,
as long as it can convert to the normal affine coordinates (which are
what is used by all ECC algorithms implemented currently, and probably
all interesting ECC algorithms in general).