Add a couple of verification tests for GOST 34.10
Generating the test vectors found yet another inane (and, of course,
undocumented) behavior in the GOST implementation included in OpenSSL;
it treats the hash inputs as little endian. Just out of curiousity, I
checked RFC 5832, which supposedly specifies this algorithm; not a
peep about endian conversions.

The more I deal with standards coming out of the CryptoPro people, the
less confidence I have in them.