Change BlockCipher::parallelism() to return the native parallelism of the implementation rather than the preferred one. Update all implementations.
Add a new function parallel_bytes() which returns
  parallelism() * BLOCK_SIZE * BUILD_TIME_CONSTANT

This is because i noticed all current calls of parallelism() just
multiplied the result by the block size already, so this simplified
that code.

The build time constant is set to 4, which was the previous default
return value of parallelism(). However the SIMD versions returned
2*native paralellism rather than 4*, so this increases the buffer
sizes used for those algorithms.

The constant multiple lives in buildh.in and build.h, and is named
BOTAN_BLOCK_CIPHER_PAR_MULT.