Improvments to the build instructions:
Note that if you want to use Python 3, you have to use 2to3. This
has come up a couple of times on the list.

Add make check to the 'for the impatient' instructions.

Mention --no-autoload.

Said that we would guess the compiler; this isn't true with the
Python version. Add examples of using ICC and Clang.

Note that prebuilt Windows binaries are sometimes available.

Talk about make install for Windows.

Fix version #s in pkg-config output.