I had this older version of algorithm_benchmark included for compatability with 1.8, but actually the signature is completely different anyway because that version took a Timer object, which doesn't exist at all anymore.
I suppose I could add an empty Timer class plus subclasses, let
someone instantiate it and pass it in, ignoring it, but I'm not
feeling this is worth the effort. It would make more sense to add a
version with this signature to 1.8, which creates a
Default_Benchmark_Timer and uses it.